Fixed sudo, edited the way how it works
thread.l now takes a table with userdata in it.
If there is not userdata, it will run as the current user.
If there is (and it should be corrent userdata, otherwise it errors), it will log in as that user everytime the task resumes and logs back to the old user after it yields or dies